requirements.txt
pandas
numpy
requests
pyarrow
fastparquet
xgboost
scikit-learn
joblib
fastapi
uvicorn
duckdb
python-dotenv
plotly
streamlit
lxml